Added tests inside Integration
